requirements.txt
Django
djangorestframework
psycopg2-binary
djangorestframework-simplejwt
python-dotenv
django-filter
drf-yasg